+++ src/osha/oira/upgrade/v1/20230726165110_enable_the_locking_feature/upgrade.py
@@ -1,0 +1,8 @@
+from ftw.upgrade import UpgradeStep
+
+
+class EnableTheLockingFeature(UpgradeStep):
+    """Enable the locking feature."""
+
+    def __call__(self):
+        self.install_upgrade_profile()